feat(dedicated): create new component for ipv4 in agoraOrder
  ref: MANAGER-13077
