+++ packages/manager/apps/dedicated/client/app/components/ip/agoraOrder/ipv4/ipv4.component.js
@@ -1,0 +1,7 @@
+import controller from './ipv4.controller';
+import template from './ipv4.html';
+
+export default {
+  controller,
+  template,
+};
